Keep v2 issuance boundaries within reviewed architecture limits
